+++ backend/accounts/transaction.go
@@ -187,6 +187,16 @@ func NewOrderedTransactions(txs []*TransactionData) OrderedTransactions {
 	return txs
 }
 
+// LatestConfirmedBalance returns the balance at the newest confirmed transaction.
+func (txs OrderedTransactions) LatestConfirmedBalance() coin.Amount {
+	for _, tx := range txs {
+		if tx.isConfirmed() {
+			return tx.Balance
+		}
+	}
+	return coin.NewAmountFromInt64(0)
+}
+
 // TimeseriesEntry contains the balance of the account at the given time.
 type TimeseriesEntry struct {
 	Time  time.Time