+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Enable Tether USDT for BTC Direct
 - Allow users to upgrade firmware during setup
 - Ethereum: improve QR code scanning and fix ERC20 QR payment requests
+- Add option for percentage on portfolio to reflect performance
 
 ## v4.51.4
 - Bundle BitBox02 and BitBox02 Nova firmware version v9.26.5